+++ qcommon/shared/q_shared.h
@@ -382,7 +382,7 @@ char *COM_Parse (char **data_p);
 char *COM_ParseExt (char **data_p, qboolean allowNewLines);
 
 void Com_sprintf (char *dest, int32_t size, char *fmt, ...);
-long Com_HashFileName (const char *fname, int32_t hashSize, qboolean sized);
+int32_t Com_HashFileName (const char *fname, int32_t hashSize, qboolean sized);
 
 void Com_PageInMemory (byte *buffer, int32_t size);
